+++ doc/changelog.md
@@ -12,6 +12,8 @@
 #### support changes
 - Fixed bug with `null` type in `tester.js`
 #### Fallen Changes
+ - Removed the `callback` option from the `button` element in `input.js`. You should use `onClick` instead
+ - Renamed the `callback` option from the `submit` element in `input.js` to `onSubmit`
  - Added new `markdown` element in `staging.js`
  - Added some error checking to the `generate` function